+++ compiler/src/org.graalvm.compiler.truffle.options/src/org/graalvm/compiler/truffle/options/PolyglotCompilerOptions.java
@@ -240,6 +240,9 @@ public final class PolyglotCompilerOptions {
     @Option(help = "Compile immediately to test Truffle compilation", category = OptionCategory.INTERNAL)
     public static final OptionKey<Boolean> CompileImmediately = new OptionKey<>(false);
 
+    @Option(help = "Compiles created call targets immediately with last tier. Disables background compilation if enabled.", category = OptionCategory.INTERNAL)
+    public static final OptionKey<Boolean> CompileAOTOnCreate = new OptionKey<>(false);
+
     @Option(help = "Enable asynchronous truffle compilation in background threads", category = OptionCategory.EXPERT)
     public static final OptionKey<Boolean> BackgroundCompilation = new OptionKey<>(true);